Add a lightweight blocking routing for user actions that may trigger navigation.
Problem
---

When Capybara clicks on an element, it immediately proceeds to the next
test assertion without actually pausing to wait for the browser to
transition to the next page.

This causes tests to become flakier because many of them are written
like this:

click_on "link on page 1"
click_on "link on page 2"
expect(page).to have_content("page 3 content")

Whereas we can add implicit wait assertions in between the first two
clicks, it is far easier to have the framework do this for us and block
if it knows the browser is transitioning.

Solution
---

The browser is to check after 100ms of clicking on an element whether it
is scheduled to transition to a new page. If not, it yields execution
back to the test context. If it is scheduled to transition, it waits for
the page to load before yielding.