Do not reveal text that is actually hidden [#18]
Previously, we would lean on `Node.textContent` to reveal a node's textual
content. However, this renders literally every text node within a DOM
node, which does not take hidden elements (like hidden DIVs or script
tags into account).

Instead, we lean on `Node.innerText`, which according to the docs
"represents the rendered text content of a node and its descendants":

https://developer.mozilla.org/en-US/docs/Web/API/Node/innerText